% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graph2sp.R
\name{arcAngle}
\alias{arcAngle}
\title{Arc of a Circle}
\usage{
arcAngle(pt1, pt2, alpha = 45, n = 100, inv = FALSE)
}
\arguments{
\item{pt1}{a numeric vector of length 2  giving the coordinates (x and y) of the first point.}

\item{alpha}{a numeric giving the value of the central angle (angular distance in degrees).}

\item{n}{the number of points to generate.}

\item{inv}{a logical. Since there are two solutions, can be used to get the second arc.}

\item{pt1}{a numeric vector of length 2  giving the coordinates (x and y) of the second point.}
}
\value{
A two-column matrix of points coordinates.
}
\description{
Compute the arc between two points with its angular distance.
}

