% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{cholera}
\alias{cholera}
\title{Cholera Data}
\format{a list of two \code{SpatialPointsDataFrame}.
\itemize{
   \item chol The location of cholera cases.
   \item pumps The location of pumps.
 }}
\source{
http://blog.rtwilson.com/john-snows-cholera-data-in-more-formats/
}
\usage{
data(cholera)
}
\description{
Cholera dataset of John Snow in London.
}
\keyword{datasets}

