% Generated by roxygen2 (4.0.1): do not edit by hand
\name{choroMap}
\alias{choroMap}
\title{Basic interactive choropleth}
\usage{
choroMap(x, col.by, breaks = 5, alpha = 1, col = NULL, popup = NULL,
  ...)
}
\arguments{
\item{x}{a \code{SpatialPolygonsDataFrame}.}

\item{col.by}{a column name of x giving values to color polygons.}

\item{breaks}{a numeric vector of cut points or a single number of intervals.}

\item{alpha}{a numeric to control the transparency.}

\item{popup}{a character vector providing popup content.}

\item{col}{a vector of color usually generated by color palettes functions.}
}
\description{
Basic interactive choropleth
}

