% Generated by roxygen2 (4.0.1): do not edit by hand
\name{contour2sp}
\alias{contour2sp}
\title{Contour Lines to SpatialLinesDataFrame}
\usage{
contour2sp(x, layer = 1, nlevels = 10)
}
\arguments{
\item{x}{a \code{SpatialGrid} object.}

\item{layer}{which layer of x to use. This may be the column name or a column number.}

\item{nlevels}{number of contour levels.}
}
\value{
a \code{SpatialLinesDataFrame}
}
\description{
This function takes a SpatialGrid object and returns
contour lines as a SpatialLinesDataFrame object.
}

