% Generated by roxygen2 (4.0.1): do not edit by hand
\name{graph2sp}
\alias{graph2sp}
\title{Convert a graph into Spatial objects}
\usage{
graph2sp(x, edges = c("lines", "gc", "arc"), alpha = 45)
}
\arguments{
\item{x}{an igraph object with at least two vertices attributes \code{"lon"} and \code{"lat"}
giving the longitude and the latitude of each vertex respectively.}

\item{edges}{a character string giving the style to use for edges (lines connecting vertices).
Can be one of \code{"lines"} for straight lines, \code{"gc"} for great circles or \code{"arc"} for user defined arcs.}

\item{alpha}{if \code{edges} is set to \code{"arc"}, a numeric giving the value of the central angle for the arcs.}
}
\value{
A list of two \code{Spatial} objects. A \code{SpatialPointsDataFrame} for the vertices
and a \code{SpatialLinesDataFrame} for the edges.
}
\description{
This function converts a graph (igraph) with geolocated vertices into Spatial objects (points and lines).
}

