% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/heatMap.R
\name{heatMap}
\alias{heatMap}
\title{Interactive heatmaps and contour lines}
\usage{
heatMap(x, bm = "stamen.toner.lite", show.contour = TRUE,
  show.gradient = TRUE, nlevels = 12, ...)
}
\arguments{
\item{x}{a \code{SpatialGridDataFrame}.}

\item{bm}{a character string giving the base map tiles server adress.
Use \code{\link[rleafmap]{bmSource}} to get a list of preconfigured servers.}

\item{show.contour}{a logical stating whether contour lines should be displayed.}

\item{show.gradient}{a logical stating whether the heatmap should be displayed.}

\item{...}{other argument to be passed to \code{\link[rleafmap]{writeMap}}}

\item{n.levels}{if \code{show.contour} is \code{TRUE}, number of contour levels.}
}
\description{
This is a wrapping function to create interactive
heatmaps and contour lines from a SpatialGridDataFrame.
}

