% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quickMap.R
\name{quickMap}
\alias{quickMap}
\title{Quickly generate an interactive map}
\usage{
quickMap(x, col.by = "", size.by = "", alpha = 1, popup = "", ...)
}
\arguments{
\item{x}{a \code{SpatialPointsDataFrame} or a \code{SpatialPolygonsDataFrame}.}

\item{col.by}{a column name of x giving colors for points or polygons.}

\item{size.by}{a column name of x giving the size for the circles.
Only relevant with \code{SpatialPointsDataFrame}}

\item{alpha}{a numeric to control the transparency.}

\item{popup}{a character vector providing popup content.}

\item{...}{other arguments to be passed to \code{\link[rleafmap]{writeMap}}}
}
\description{
This function can be used to quickly generate interactive maps
from a \code{SpatialPointsDataFrame} or a \code{SpatialPolygonsDataFrame}.
}
\examples{
\dontrun{
data(cholera)
quickMap(cholera$deaths, col.by = "Count", size.by = "Count")
}
}

