% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epimap-package.R
\docType{data}
\name{cholera}
\alias{cholera}
\title{Cholera Data}
\format{a list of two \code{SpatialPointsDataFrame}.
\itemize{
   \item deaths Cholera Death locations.
   \item deaths.den Cholera deaths density.
   \item pumps Pumps locations.
 }}
\source{
http://blog.rtwilson.com/john-snows-cholera-data-in-more-formats/
}
\usage{
data(cholera)
}
\description{
Cholera dataset of John Snow in London.
}
\keyword{datasets}

