% Generated by roxygen2 (4.0.1): do not edit by hand
\name{netMap}
\alias{netMap}
\title{Interactive network on a map}
\usage{
netMap(x, v.size = 5, v.col = 2, bm = "cartodb.darkmatter.nolab", ...)
}
\arguments{
\item{x}{an \code{igraph} object with at least two vertices attributes \code{"lon"} and \code{"lat"}
giving the longitude and the latitude of each vertex respectively.}

\item{bm}{a character string giving the base map tiles server adress.
Use \code{\link[rleafmap]{bmSource}} to get a list of preconfigured servers.}
}
\description{
This is a wrapping function to create interactive
network on a map from an \code{igraph} object.
}

