% Generated by roxygen2 (4.0.1): do not edit by hand
\name{quickMap}
\alias{quickMap}
\title{Quickly generate an interactive map}
\usage{
quickMap(x, col.by = "", size.by = "", alpha = 1, popup = "", ...)
}
\arguments{
\item{x}{a \code{SpatialPointsDataFrame} or a \code{SpatialPolygonsDataFrame}.}

\item{col.by}{a column name of x giving colors for points or polygons.}

\item{size.by}{a column name of x giving the size for the circles.
Only relevant with \code{SpatialPointsDataFrame}}

\item{alpha}{a numeric to control the transparency.}

\item{popup}{a character vector providing popup content.}
}
\description{
Quickly generate an interactive map
}

